+++ contrib/PaddleRendering/pprndr/models/layers/mlps.py
@@ -13,12 +13,11 @@
 #  limitations under the License.
 
 import math
-from typing import List, Tuple, Union
+from typing import List, Tuple, Union, Callable, Any
 
 import paddle
 import paddle.nn as nn
 import paddle.nn.functional as F
-import numpy as np
 from pprndr.apis import manager
 
 try:
@@ -53,97 +52,6 @@ PADDLE_ACTIVATION = {
     F.softplus,
     "none":
     None
-}
-
-
-class GeometricInit(object):
-    def __init__(self, params):
-        assert (len(params) == 2)
-        self.bias = float(params[0])
-        self.multires = bool(params[1])
-
-    def initialize(self,
-                   layers: list,
-                   skip_layers: list = None,
-                   dims: list = None):
-        assert (len(layers) > 1)
-        for i, layer in enumerate(layers):
-            if i == len(layers) - 1:
-                dim = dims[i]  # input_dim of the last layer
-                which_layer = "last"
-
-            elif i == 0:
-                dim = dims[i + 1]  # output_dim of the first layer
-                which_layer = "first"
-
-            elif i in skip_layers:
-                dim = [dims[i + 1], (dims[0] - 3)]
-                which_layer = "skip"
-
-            else:
-                dim = dims[i + 1]
-                which_layer = "else"
-
-            self.initialize_layer(layer, dim, which_layer)
-
-    def initialize_layer(self, layer, dim, which_layer):
-        assert (which_layer in ["first", "skip", "last", "else"])
-        if which_layer == "last":
-            # Init. weight
-            _weight = layer.weight.numpy()
-            _weight = np.random.normal(
-                np.sqrt(np.pi) / np.sqrt(dim), 0.0001, _weight.shape)
-            nn.initializer.Assign(_weight)(layer.weight)
-            if layer.bias is not None:
-                # Init bias
-                _bias = layer.bias.numpy()
-                _bias[...] = -self.bias
-                nn.initializer.Assign(_bias)(layer.bias)
-
-        elif which_layer == "first" and self.multires:
-            # Init. weight
-            # Note dim should be output dim of this layer
-            _weight = layer.weight.numpy()
-            _weight = np.random.normal(0.0,
-                                       np.sqrt(2) / np.sqrt(dim), _weight.shape)
-            _weight[3:, :] = 0.0
-
-            nn.initializer.Assign(_weight)(layer.weight)
-            if layer.bias is not None:
-                # Init bias
-                _bias = layer.bias.numpy()
-                _bias[...] = 0
-                nn.initializer.Assign(_bias)(layer.bias)
-
-        elif which_layer == "skip" and self.multires:
-            out_dim = dim[0]
-            embed_dim = dim[1]
-            # Init weight
-            _weight = layer.weight.numpy()
-            _weight = np.random.normal(0.0,
-                                       np.sqrt(2) / np.sqrt(out_dim),
-                                       _weight.shape)
-            _weight[-embed_dim:] = 0.0
-            nn.initializer.Assign(_weight)(layer.weight)
-            if layer.bias is not None:
-                # Init bias
-                _bias = layer.bias.numpy()
-                _bias[...] = 0.0
-                nn.initializer.Assign(_bias)(layer.bias)
-        else:
-            _weight = layer.weight.numpy()
-            _weight = np.random.normal(0.0,
-                                       np.sqrt(2) / np.sqrt(dim), _weight.shape)
-            nn.initializer.Assign(_weight)(layer.weight)
-            if layer.bias is not None:
-                # Init bias
-                _bias = layer.bias.numpy()
-                _bias[...] = 0.0
-                nn.initializer.Assign(_bias)(layer.bias)
-
-
-PADDLE_INITIALIZATION = {
-    "geometric_init": GeometricInit,
 }
 
 
@@ -240,8 +148,8 @@ class MLP(nn.Layer):
                  skip_connection_scale: float = None,
                  with_bias: bool = True,
                  activation: Union[str, nn.Layer] = "relu",
-                 output_activation: str = "none",
-                 weight_init: list = None,
+                 output_activation: Union[str, nn.Layer] = "none",
+                 weight_init: Any = None,
                  weight_normalization: bool = False):
         super(MLP, self).__init__()
 
@@ -272,41 +180,33 @@ class MLP(nn.Layer):
 
         self.hidden_dim = hidden_dim
         self.skip_layers = skip_layers if skip_layers is not None else set()
-        if len(list(
-                self.skip_layers)) > 0 and skip_connection_way == "concat_in":
-            for skip_layer in list(self.skip_layers):
-                hid = self.hidden_dim[skip_layer - 1] - self.input_dim
-                self.hidden_dim[skip_layer - 1] = hid
+        if len(self.skip_layers) > 0 and skip_connection_way == "concat_in":
+            for skip_layer in self.skip_layers:
+                self.hidden_dim[skip_layer - 1] -= self.input_dim
 
-        if isinstance(activation, str):
-            self.activation = PADDLE_ACTIVATION[activation.lower()]
-        elif isinstance(activation, nn.Layer):
-            self.activation = activation
-        else:
-            raise NotImplementedError(
-                "activation should be either str or nn.Layer.")
-
-        self.output_activation = PADDLE_ACTIVATION[output_activation.lower()]
+        self.activation = self._populate_activation(activation)
+        self.output_activation = self._populate_activation(output_activation)
         self.layers = self._populate_layers()
 
         # Weight init_fn.
         if weight_init is not None:
-            init_fn_name = weight_init[0].lower()
-            self.weight_init_fn = PADDLE_INITIALIZATION[init_fn_name](
-                weight_init[1:])
-            if init_fn_name == "geometric_init":
-                assert (len(self.layers) > 1)
-                self.weight_init_fn.initialize(
-                    self.layers, self.skip_layers,
-                    [self.input_dim] + self.hidden_dim)
-            else:
-                pass
+            weight_init.initialize(self.layers, self.skip_layers,
+                                   [self.input_dim] + self.hidden_dim)
 
-        # Weight normlaization
-        if weight_normalization is not None:
-            if weight_normalization:
-                for layer in self.layers:
-                    nn.utils.weight_norm(layer, dim=1)
+        # Weight normalization
+        if weight_normalization:
+            for layer in self.layers:
+                nn.utils.weight_norm(layer, dim=1)
+
+    def _populate_activation(self,
+                             activation: Union[str, nn.Layer]) -> Callable:
+        if isinstance(activation, str):
+            activation = PADDLE_ACTIVATION[activation.lower()]
+        elif not isinstance(activation, nn.Layer):
+            raise NotImplementedError(
+                "activation should be either str or nn.Layer.")
+
+        return activation
 
     def _populate_layers(self) -> nn.LayerList:
         layers = nn.LayerList()
